refactor: move SqlHelper into DatabaseService
